requirements.txt
azure-storage-blob==12.14.1
tenacity==8.2.2
pytest==8.3.3